requirements-pyqt.txt
PyQt5>=5.15
pyqtgraph>=0.13
pyserial>=3.5
numpy>=1.20
PyOpenGL>=3.1
PyOpenGL_accelerate>=3.1
